assessment/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==2.0.7
django-business-rules
psycopg2-binary>=2.8
djangorestframework


